Major improvements to swing and stab detection
From early playtests it was quite obvious that even though the tuned parameters worked for our body and arm lengths, it wasn't really compatible with other body types.

It now uses acceleration data to be able to detect swings earlier and in a much more generic fashion. It's also more lenient.